% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Create a dotplot of phosphopeptide intensity for all donors - Proxy to Study Package}
\usage{
dotplot(idmult, study)
}
\arguments{
\item{idmult}{The idmult of the phosphosite}

\item{study}{The study}
}
\value{
A dotplot
}
\description{
This function returns intensity values for each donor time 0. Data points are indicated as observed or imputed.
}
\examples{
dotplot(idmult = "3618_1", study="***REMOVED***")
}
