% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentNames.R
\name{EnrichmentNames}
\alias{EnrichmentNames}
\title{Parse Inference.Results to obtain names of objects within - Proxy to Study Package}
\usage{
EnrichmentNames(study)
}
\arguments{
\item{study}{The study}
}
\value{
json
}
\description{
This function returns a json object.
}
\examples{
 EnrichmentNames("***REMOVED***")
}
