% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentResults.R
\name{getEnrichmentResults}
\alias{getEnrichmentResults}
\title{Search Inference.Results for a particular dataset - Proxy to Study Package}
\usage{
getEnrichmentResults(testCategory, annotation, study)
}
\arguments{
\item{testCategory}{The test category}

\item{annotation}{The annotation}

\item{study}{The study}
}
\value{
a table
}
\description{
This function returns a table of results.
}
\examples{
 getEnrichmentResults("DonorDifferentialPhosphorylation","GSEA_biocarta_modelI","***REMOVED***")
}
