% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInferenceResults.R
\name{getInferenceResults}
\alias{getInferenceResults}
\title{Search Inference.Results for a particular dataset - Proxy to Study Package}
\usage{
getInferenceResults(testCategory, test, study)
}
\arguments{
\item{testCategory}{The test category}

\item{test}{The test}

\item{study}{The study}
}
\value{
a table
}
\description{
This function returns a table of results.
}
\examples{
 getInferenceResults("DonorDifferentialPhosphorylation","Donor1vDonor2","***REMOVED***")
}
