% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceNames.R
\name{inferenceNames}
\alias{inferenceNames}
\title{Parse Inference.Results to obtain names of objects within - Proxy to Study Package}
\usage{
inferenceNames(study)
}
\arguments{
\item{study}{The study}
}
\value{
json
}
\description{
This function returns a json object.
}
\examples{
 inferenceNames("***REMOVED***")
}
