% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\name{lineplot}
\alias{lineplot}
\title{This function returns a plot of line segments connecting the medians for each strain across the timecourse. Data points are indicated as observed or imputed. 0 time points are called 'baseline'. - Proxy to Study Package}
\usage{
lineplot(idmult, study)
}
\arguments{
\item{idmult}{The idmult of the phosphosite}

\item{study}{The study}
}
\value{
A plot
}
\description{
This function returns a plot of line segments connecting the medians for each strain across the timecourse. Data points are indicated as observed or imputed. 0 time points are called 'baseline'. - Proxy to Study Package
}
\examples{
lineplot(idmult = "7072_2", study="***REMOVED***")
}
