% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphodotplot.R
\name{phosphodotplot}
\alias{phosphodotplot}
\title{Create a dotplot of phosphopeptide intensity for all donors - Proxy to Study Package}
\usage{
phosphodotplot(idmult, study)
}
\arguments{
\item{idmult}{The id of the protein group}

\item{study}{The study}
}
\value{
A dotplot
}
\description{
This function returns intensity values for each donor. Data points are indicated as observed or imputed.
}
\examples{
phosphodotplot(idmult = "9382_2", study="***REMOVED***plots")
}
