% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteindata.R
\name{proteindata}
\alias{proteindata}
\title{#' Return protein metadata -  Proxy to Study Package}
\usage{
proteindata(id, study)
}
\arguments{
\item{id}{The id of the protein group}

\item{study}{The study}
}
\value{
A data frame that includes lots of goodies!
}
\description{
This function returns the MS metadata for protein groups
}
\details{
This function hits the evidence.txt file.
}
\examples{
proteindata(id = "1825")
}
