% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteindotplot.R
\name{proteindotplot}
\alias{proteindotplot}
\title{Create a dotplot of protein group intensity for all donors - Proxy to Study Package}
\usage{
proteindotplot(idmult, study)
}
\arguments{
\item{study}{The study}

\item{id}{The id of the protein group}
}
\value{
A dotplot
}
\description{
This function returns intensity values for each donor time 0. Data points are indicated as observed or imputed.
}
\examples{
proteindotplot(id = "1825", study="***REMOVED***plots")
}
