% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitedata.R
\name{sitedata}
\alias{sitedata}
\title{Return phosphopeptide metadata for a given phosphosite_multiplicity summation - Proxy to Study Package}
\usage{
sitedata(idmult, study)
}
\arguments{
\item{idmult}{The idmult of the phosphosite}

\item{study}{The study}
}
\value{
A data frame that includes localized phosphopeptides and the experiments in which each phosphopeptide was observed (or matched). Protein_Site (HGNC symbol), Modifications, Modified sequence, Type, Experiment (semicolon delimited), Obervations (n), Median Intensity (raw), Proteins,	Leading proteins,	Leading razor protein
}
\description{
This function returns the phosphopeptides (with site localization and other metadata) used to quantify the levels of a given phosphosite.
}
\details{
This function hits the evidence.txt file. For singly phsphorylated peptides the phosphopeptide with the most quantifications is used. For multiply phosphorylated peptides the union of all phosphopeptides is used for the quantification. It is unclear what rule, if any is used by MQ to assign a given multiply phosphorylated peptide when multiple multiply phosphorylated peptides are observed in the same sample.
}
\examples{
sitedata(idmult = "221_1", study="PLA2G2plots")
}
