% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineplot.R
\name{splineplot}
\alias{splineplot}
\title{Plot the fitted splines for a phosphopeptide for all compounds - Proxy to Study Package}
\usage{
splineplot(idmult, study)
}
\arguments{
\item{idmult}{The idmult of the phosphosite}

\item{study}{The study}
}
\value{
A plot
}
\description{
Plot the fitted splines for a phosphopeptide for all compounds - Proxy to Study Package
}
\examples{
splineplot(idmult = "221_1", study="PLA2G2plots")
}
