% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StimStrainDotplot.R
\name{StimStrainDotplot}
\alias{StimStrainDotplot}
\title{Create a dotplot of phosphopeptide intensity for all donors - Proxy to Study Package}
\usage{
StimStrainDotplot(idmult, study)
}
\arguments{
\item{idmult}{The idmult of the phosphosite}

\item{study}{The study}
}
\value{
A dotplot
}
\description{
This function returns intensity values for each donor time 0. Data points are indicated as observed or imputed.
}
\examples{
StimStrainDotplot(idmult = "6173_1", study="***REMOVED***")
}
