% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrainStimDotplot.R
\name{StrainStimDotplot}
\alias{StrainStimDotplot}
\title{Create a dotplot of phosphopeptide intensity for all donors - Proxy to Study Package}
\usage{
StrainStimDotplot(idmult, study)
}
\arguments{
\item{idmult}{The idmult of the phosphosite}

\item{study}{The study}
}
\value{
A dotplot
}
\description{
This function returns intensity values for each donor time 0. Data points are indicated as observed or imputed.
}
\examples{
StrainStimDotplot(idmult = "6173_1")
}
